backend/requirements/common.txt
alembic==1.4.3
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
fastapi==0.61.2
h11==0.11.0
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pydantic==1.7.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.12.2
